Rename semanticNonNull to nullability
